requirements.txt
aenum==3.1.15
altgraph==0.17
amqp==2.5.2
asgiref==3.6.0
astroid==2.4.2
attrs==19.3.0
audioread==2.1.8
awsebcli==3.20.3
backcall==0.1.0
beautifulsoup4==4.9.3
billiard==3.6.3.0
bleach==3.1.1
boto3==1.26.131
botocore==1.29.131
branca==0.6.0
cachetools==5.3.0
cairocffi==1.2.0
CairoSVG==2.5.1
celery==4.4.1
cement==2.8.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.3
configparser==5.0.0
coverage==5.5
cssselect2==0.4.1
cursor==1.1.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
deprecation==2.1.0
diff-match-patch==20200713
Django==3.2.19
django-choices==1.7.2
django-ckeditor==5.9.0
django-colorfield==0.8.0
-e git+https://github.com/APSL/django-cookie-law.git@6d5225f53db13105d646d227653dc5767c842f04#egg=django_cookie_law
django-filter==2.3.0
django-import-export==2.5.0
django-js-asset==1.2.2
django-multiselectfield==0.1.12
django-recaptcha==3.0.0
django-storages==1.11.1
django-tempus-dominus==5.1.2.13
docopt==0.6.2
entrypoints==0.3
et-xmlfile==1.0.1
eyed3==0.9.7
facebook-sdk==3.1.0
filetype==1.2.0
Fiona==1.9.4.post1
folium==0.14.0
future==0.16.0
geopandas==0.10.2
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
greenlet==0.4.15
h5py==2.10.0
html5lib==1.1
httplib2==0.18.1
humanize==3.0.0
idna==2.9
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.16.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.0.0
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.1.0
kombu==4.6.8
lazy-object-proxy==1.4.3
librosa==0.7.2
llvmlite==0.31.0
lml==0.1.0
lxml==4.9.3
MarkupPy==1.14
MarkupSafe==1.1.1
matplotlib==3.2.0
mccabe==0.6.1
mistune==0.8.4
mutagen==1.45.1
mypy-extensions==0.4.3
mysqlclient==1.4.6
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
num2words==0.5.10
numba==0.48.0
numpy==1.18.1
oauthlib==3.1.1
odfpy==1.4.1
opencv-python==4.5.4.60
openpyxl==3.0.9
packaging==23.1
pandas==1.3.5
pandocfilters==1.4.2
parso==0.6.2
pathspec==0.9.0
PDFNetPython3==9.2.0
pefile==2023.2.7
pickleshare==0.7.5
Pillow==9.5.0
prometheus-client==0.7.1
prompt-toolkit==3.0.4
protobuf==4.23.0
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.20
pydub==0.23.1
pyexcel==0.6.6
pyexcel-io==0.6.4
pyexcel-webio==0.1.4
pygame==2.1.0
Pygments==2.6.1
pyinstaller==5.10.1
pyinstaller-hooks-contrib==2023.2
PyJWT==1.7.1
pylint==2.6.0
pymssql==2.2.7
PyNaCl==1.4.0
pyparsing==2.4.6
PyPDF2==1.26.0
Pyphen==0.10.0
pypiwin32==223
pyproj==3.2.1
PyQt5==5.15.9
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.0
pyqt5-tools==5.15.4.3.2
PyQt6==6.1.0
pyqt6-plugins==6.1.0.2.2
PyQt6-Qt6==6.3.1
PyQt6-sip==13.4.0
pyqt6-tools==6.1.0.3.2
pyqtgraph==0.12.1
pyrsistent==0.15.7
PySide2==5.15.2.1
PySide6==6.3.2
PySide6-Addons==6.3.2
PySide6-Essentials==6.3.2
PySocks==1.7.1
pytesseract==0.3.8
python-barcode==0.13.1
python-dateutil==2.8.1
python-docx==0.8.11
python-dotenv==0.21.0
python-Levenshtein==0.12.2
pytz==2019.3
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==19.0.0
qrcode==6.1
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
qt6-applications==6.1.0.2.2
qt6-tools==6.1.0.1.2
qtconsole==4.7.1
QtPy==1.9.0
redis==3.4.1
reportlab==3.5.62
reportlab-qrcode==0.1.4
requests==2.26.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.9
s3transfer==0.6.1
scikit-learn==0.22.2.post1
scipy==1.4.1
semantic-version==2.8.5
Send2Trash==1.5.0
sentry-sdk==0.14.3
shapely==2.0.1
shiboken2==5.15.2.1
shiboken6==6.3.2
six==1.14.0
soundfile==0.12.1
soupsieve==2.2
sqlparse==0.3.0
tablib==3.0.0
telnyx==1.1.0
tempus==1.1.0
tenacity==8.2.3
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
texttable==1.6.4
tinycss2==1.1.0
toml==0.10.2
tornado==6.0.4
traitlets==4.3.3
tweepy==4.4.0
twilio==6.45.4
typed-ast==1.4.1
typing-extensions==3.7.4.1
uritemplate==4.1.1
urllib3==1.26.12
vine==1.3.0
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.11.2
xhtml2pdf==0.2.4
xlrd==2.0.1
xlwt==1.3.0
zipp==3.1.0
